+++ src/useSubmission.ts
@@ -1,6 +1,6 @@
-import { useCallback, SyntheticEvent } from 'react';
-import { hasValue } from './useGenericValues';
-import { Errors } from './useErrors';
+import {
+  useCallback, SyntheticEvent, useState, useEffect, useMemo,
+} from 'react';
 import { useInvokeCount, useInvoking } from './useFunctionUtils';
 
 export interface SubmissionHandler {
@@ -11,7 +11,12 @@ export interface SubmitHandler {
 }
 export interface UseSubmissionProps {
   readonly onSubmit: SubmissionHandler;
-  readonly validator: () => Promise<Errors> | Errors;
+  readonly validator?: () => Promise<void> | void;
+  /**
+   * Determines if submission should be disabled. Generally,
+   * you want to disable if there are errors.
+   */
+  readonly disabled?: boolean;
 }
 
 export interface UseSubmissionHook {
@@ -51,19 +56,39 @@ export interface UseSubmissionHook {
  *     onSubmit, validator
  *   });
  */
-export function useSubmission({ validator, onSubmit }: UseSubmissionProps): UseSubmissionHook {
+export function useSubmission({
+  onSubmit,
+  validator,
+  disabled = false,
+}: UseSubmissionProps): UseSubmissionHook {
+  const [isReadyToSubmit, setIsReadyToSubmit] = useState(false);
   // track submission count
-  const [wrappedOnSubmit, submitCount] = useInvokeCount(onSubmit);
-  const rawSubmissionHandler = useCallback(async (event?: SyntheticEvent): Promise<void> => {
+  const [submitWithInvokeCount, submitCount] = useInvokeCount(onSubmit);
+  const [submitWithInvokingTracker, isSubmitting] = useInvoking(submitWithInvokeCount);
+  const validationFnc = useMemo(() => validator || ((): void => undefined), [validator]);
+  const [validate, isValidating] = useInvoking(validationFnc);
+
+  // track when to kick off submission
+  useEffect(() => {
+    if (isReadyToSubmit && !isValidating) {
+      setIsReadyToSubmit(false);
+      if (!disabled) {
+        submitWithInvokingTracker();
+      }
+    }
+  }, [
+    disabled,
+    isValidating,
+    isReadyToSubmit,
+    setIsReadyToSubmit,
+    submitWithInvokingTracker,
+  ]);
+  const submit = useCallback((event?: SyntheticEvent) => {
     if (event) {
       event.preventDefault();
     }
-    const errors = await validator();
-    if (!hasValue(errors)) {
-      await wrappedOnSubmit();
-    }
-  }, [validator, wrappedOnSubmit]);
-  // track if is submitting
-  const [submit, isSubmitting] = useInvoking(rawSubmissionHandler);
+    validate();
+    setIsReadyToSubmit(true);
+  }, [setIsReadyToSubmit, validate]);
   return { isSubmitting, submitCount, submit };
 }