+++ test/useSubmission.test.ts
@@ -1,55 +1,39 @@
 import { renderHook, act } from '@testing-library/react-hooks';
 import { useSubmission } from '../src';
+import { UseSubmissionProps, UseSubmissionHook } from '../src/useSubmission';
 
+const sleep = (duration: number) => {
+return new Promise<void>((res) => {
+    setTimeout(() => {
+        res();
+    }, duration * 1000);
+});
+};
 describe("useSubmission", () => {
     it.todo("doesn't allow invalid arguments");
-    it("supports async validators", async () => {
-        const onSubmit = jest.fn(() => { });
-        let { result, waitForNextUpdate } = renderHook(() => useSubmission({
-            validator: async () => {
-                return { name: "test is an error" };
-            },
-            onSubmit,
-        }));
+    it.todo("supports async submit handlers");
+    it("only submits if disabled is set to false", async () => {
+        const onSubmit = jest.fn();
+        const { result, rerender } = renderHook<UseSubmissionProps, UseSubmissionHook>(
+            // @ts-ignore
+            ({ disabled } = {}) => useSubmission({
+                onSubmit,
+                disabled,
+            })
+        );
         act(() => {
             result.current.submit();
         });
-        await waitForNextUpdate();
-        expect(result.current.submitCount).toBe(0);
-        expect(onSubmit.mock.calls.length).toBe(0);
-        ({ result, waitForNextUpdate } = renderHook(() => useSubmission({
-            validator: async () => {
-                return {};
-            },
-            onSubmit,
-        })));
-        act(() => {
-            result.current.submit();
-        });
-        await waitForNextUpdate();
         expect(result.current.submitCount).toBe(1);
         expect(onSubmit.mock.calls.length).toBe(1);
-    });
-    it.todo("supports async submit handlers");
-    it("only submits after the form is error free", async () => {
-        const onSubmit = jest.fn(() => { });
-        const { result, waitForNextUpdate } = renderHook(() => useSubmission({
-            validator: () => ({ name: "this is an error" }),
-            onSubmit,
-        }));
-        act(() => {
-            result.current.submit();
-            result.current.submit();
-            result.current.submit();
-            result.current.submit();
-        });
-        await waitForNextUpdate();
-        expect(result.current.submitCount).toBe(0);
+
+        onSubmit.mockReset();
+        rerender({ disabled: true, onSubmit });
+        expect(result.current.submitCount).toBe(1);
         expect(onSubmit.mock.calls.length).toBe(0);
     });
     it("calls preventDefault when submitting", async () => {
-        const { result, waitForNextUpdate } = renderHook(() => useSubmission({
-            validator: () => ({}),
+        const { result } = renderHook(() => useSubmission({
             onSubmit: () => { },
         }));
         const preventDefault = jest.fn(() => { });
@@ -57,28 +41,26 @@ describe("useSubmission", () => {
             // @ts-ignore
             result.current.submit({ preventDefault });
         });
-        await waitForNextUpdate();
         expect(preventDefault.mock.calls.length).toBe(1);
     });
     it("supports a submission count", async () => {
-        const { result, waitForNextUpdate } = renderHook(() => useSubmission({
-            validator: () => ({}),
+        const { result } = renderHook(() => useSubmission({
             onSubmit: () => { },
         }));
         act(() => {
             result.current.submit();
-            result.current.submit();
-            result.current.submit();
+        });
+        act(() => {
             result.current.submit();
         });
-        await waitForNextUpdate();
-        expect(result.current.submitCount).toBe(4);
+        act(() => {
+            result.current.submit();
+        });
+        expect(result.current.submitCount).toBe(3);
     });
     it("determines when submission is happening", async () => {
         let { result, waitForNextUpdate } = renderHook(() => useSubmission({
-            // @ts-ignore
-            validator: () => ({}),
-            onSubmit: () => {},
+            onSubmit: () => sleep(0.250),
         }));
         expect(result.current.isSubmitting).toBe(false);
         act(() => {